+++ src/components/menu.py
@@ -3,6 +3,19 @@ from lib.sparql_queries import list_graphs
 from lib.utils import load_config
 
 
+def on_graph_selection():
+    """Callback function for the graph selection."""
+
+    # Compute all graph labels, and fetch radio button value
+    graphs_labels = [graph['label'] for graph in st.session_state["all_graphs"]]
+    graph_label = st.session_state['radio-btn-graph-selection']
+
+    # Find the selected graph, and set the session variable
+    index = graphs_labels.index(graph_label)
+    st.session_state['activated_graph_index'] = index
+
+
+
 def menu() -> None:
     """Component function: the sidebar"""
 
@@ -61,6 +74,7 @@ def menu() -> None:
                 st.cache_data.clear()
                 st.cache_resource.clear()
                 st.rerun()
+                
 
             # Fetch all graphs from the endpoint, and add the default one (at first position)
             if "all_graphs" not in st.session_state:
@@ -69,8 +83,7 @@ def menu() -> None:
                 st.session_state["all_graphs"] = [{
                     'uri': None,
                     'label': 'Default',
-                    'comment': "No comment",
-                    'activated': True
+                    'comment': "No comment"
                 }]
 
                 # Fetch and add other graphs
@@ -79,19 +92,21 @@ def menu() -> None:
                     st.session_state["all_graphs"].append({
                         'uri': graph['uri'],
                         'label': graph['label'],
-                        'comment': graph['comment'],
-                        'activated': False
+                        'comment': graph['comment']
                     })
+                
+                st.session_state['activated_graph_index'] = 0
 
-            # Allow user to activate/deactivate a graph
-            st.sidebar.markdown('#### Endpoint graph List:')
-            for i, graph in enumerate(st.session_state["all_graphs"]):
-                # Make the session match the GUI
-                if st.sidebar.checkbox(graph['label'], value=graph['activated'], key=f'graph-checkbox-{i}'):
-                    if st.session_state['all_graphs'][i]['activated'] != True:
-                        st.session_state['all_graphs'][i]['activated'] = True
-                        st.rerun()
-                else:
-                    if st.session_state['all_graphs'][i]['activated'] != False:
-                        st.session_state['all_graphs'][i]['activated'] = False
-                        st.rerun()
+            st.sidebar.text('')
+
+            # Allow user to choose the graph to activate
+            graphs_labels = [graph['label'] for graph in st.session_state["all_graphs"]]
+
+            # Manage the graph selection
+            st.sidebar.radio('Selected graph', options=graphs_labels, index=st.session_state['activated_graph_index'], key='radio-btn-graph-selection', on_change=on_graph_selection)
+
+            # activated_label = st.sidebar.radio('Selected graph', options=graphs_labels, index=st.session_state['activated_graph_index'], key='graph-radio', on_change=on_graph_selection)
+            # if activated_label:
+            #     index = graphs_labels.index(activated_label)
+            #     if index != st.session_state['activated_graph_index']:
+            #         st.session_state['activated_graph_index'] = index